requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
pydantic==1.9.0
requests==2.26.0
python-dotenv==0.19.0